crud/requirements.txt
python-dotenv
psycopg2-binary
sqlmodel
fastapi
uvicorn
pydantic[email]
passlib[bcrypt]



